Arranged editor methods from Note to Editor class
